pyproject.toml
[tool.poetry]
name = "xkcd_pipelines"
version = "0.1.0"
description = "This pipeline will be fetching, transforming, and loading XKCD comic data."
authors = ["Daniel Ibarbo Betancur <danielibarbo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
requests = "^2.32.3"
psycopg2 = "^2.9.10"
apache-airflow = "^2.10.5"
psycopg2-binary = "^2.9.10"
pendulum = "^3.0.0"
dotenv = "^0.9.9"
mocker = "^1.1.1"
pytest = "^8.3.5"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
pytest = "^8.3.5"
black = "^23.9.1"
pytest-mock = "^3.0.0"
types-requests = "^2.32.0.20250328"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
